requirements.txt
streamlit
rdkit-pypi
scikit-learn
xgboost
joblib
pandas
numpy